EF upgrade, multiple selection started
git-svn-id: https://subversion.assembla.com/svn/productlist/trunk@65 5957f8c8-545f-48dc-b781-7e43294cae7e

+++ Hotel.cs
@@ -16,8 +16,8 @@ namespace ProductList
     {
         public Nullable<int> LocationId { get; set; }
         public Nullable<int> Rooms { get; set; }
-        public Nullable<int> Star { get; set; }
-        public Nullable<int> CustomerRating { get; set; }
+        public Nullable<decimal> Star { get; set; }
+        public Nullable<decimal> CustomerRating { get; set; }
         public string Address { get; set; }
         public string CurrencyCode { get; set; }
